requirements.txt
schema==0.6
pyyaml==3.13
jinja2==2.10